+++ src/system/components/Badge.svelte
@@ -67,7 +67,7 @@
     --badge-primary-text-font-family: var(--font-sans);
     --badge-primary-text-font-size: var(--font-size-md);
     --badge-primary-text-font-weight: var(--font-weight-normal);
-    --badge-primary-text-line-height: var(--line-height-xs);
+    --badge-primary-text-line-height: var(--line-height-none);
     --badge-primary-border-width: var(--border-width-1);
     --badge-primary-radius: var(--radius-full);
     --badge-primary-padding: var(--space-6);
@@ -82,7 +82,7 @@
     --badge-accent-text-font-family: var(--font-sans);
     --badge-accent-text-font-size: var(--font-size-md);
     --badge-accent-text-font-weight: var(--font-weight-normal);
-    --badge-accent-text-line-height: var(--line-height-xs);
+    --badge-accent-text-line-height: var(--line-height-none);
     --badge-accent-border-width: var(--border-width-1);
     --badge-accent-radius: var(--radius-full);
     --badge-accent-padding: var(--space-6);
@@ -97,7 +97,7 @@
     --badge-neutral-text-font-family: var(--font-sans);
     --badge-neutral-text-font-size: var(--font-size-md);
     --badge-neutral-text-font-weight: var(--font-weight-normal);
-    --badge-neutral-text-line-height: var(--line-height-xs);
+    --badge-neutral-text-line-height: var(--line-height-none);
     --badge-neutral-border-width: var(--border-width-1);
     --badge-neutral-radius: var(--radius-full);
     --badge-neutral-padding: var(--space-6);
@@ -112,7 +112,7 @@
     --badge-alternate-text-font-family: var(--font-sans);
     --badge-alternate-text-font-size: var(--font-size-md);
     --badge-alternate-text-font-weight: var(--font-weight-normal);
-    --badge-alternate-text-line-height: var(--line-height-xs);
+    --badge-alternate-text-line-height: var(--line-height-none);
     --badge-alternate-border-width: var(--border-width-1);
     --badge-alternate-radius: var(--radius-full);
     --badge-alternate-padding: var(--space-6);
@@ -127,7 +127,7 @@
     --badge-canvas-text-font-family: var(--font-sans);
     --badge-canvas-text-font-size: var(--font-size-md);
     --badge-canvas-text-font-weight: var(--font-weight-normal);
-    --badge-canvas-text-line-height: var(--line-height-xs);
+    --badge-canvas-text-line-height: var(--line-height-none);
     --badge-canvas-border-width: var(--border-width-1);
     --badge-canvas-radius: var(--radius-full);
     --badge-canvas-padding: var(--space-6);
@@ -142,7 +142,7 @@
     --badge-special-text-font-family: var(--font-sans);
     --badge-special-text-font-size: var(--font-size-md);
     --badge-special-text-font-weight: var(--font-weight-normal);
-    --badge-special-text-line-height: var(--line-height-xs);
+    --badge-special-text-line-height: var(--line-height-none);
     --badge-special-border-width: var(--border-width-1);
     --badge-special-radius: var(--radius-full);
     --badge-special-padding: var(--space-6);
@@ -157,7 +157,7 @@
     --badge-success-text-font-family: var(--font-sans);
     --badge-success-text-font-size: var(--font-size-md);
     --badge-success-text-font-weight: var(--font-weight-normal);
-    --badge-success-text-line-height: var(--line-height-xs);
+    --badge-success-text-line-height: var(--line-height-none);
     --badge-success-border-width: var(--border-width-1);
     --badge-success-radius: var(--radius-full);
     --badge-success-padding: var(--space-6);
@@ -172,7 +172,7 @@
     --badge-warning-text-font-family: var(--font-sans);
     --badge-warning-text-font-size: var(--font-size-md);
     --badge-warning-text-font-weight: var(--font-weight-normal);
-    --badge-warning-text-line-height: var(--line-height-xs);
+    --badge-warning-text-line-height: var(--line-height-none);
     --badge-warning-border-width: var(--border-width-1);
     --badge-warning-radius: var(--radius-full);
     --badge-warning-padding: var(--space-6);
@@ -187,7 +187,7 @@
     --badge-danger-text-font-family: var(--font-sans);
     --badge-danger-text-font-size: var(--font-size-md);
     --badge-danger-text-font-weight: var(--font-weight-normal);
-    --badge-danger-text-line-height: var(--line-height-xs);
+    --badge-danger-text-line-height: var(--line-height-none);
     --badge-danger-border-width: var(--border-width-1);
     --badge-danger-radius: var(--radius-full);
     --badge-danger-padding: var(--space-6);
@@ -202,7 +202,7 @@
     --badge-info-text-font-family: var(--font-sans);
     --badge-info-text-font-size: var(--font-size-md);
     --badge-info-text-font-weight: var(--font-weight-normal);
-    --badge-info-text-line-height: var(--line-height-xs);
+    --badge-info-text-line-height: var(--line-height-none);
     --badge-info-border-width: var(--border-width-1);
     --badge-info-radius: var(--radius-full);
     --badge-info-padding: var(--space-6);
@@ -215,7 +215,7 @@
     display: inline-flex;
     align-items: center;
     gap: var(--space-6);
-    line-height: var(--line-height-xs);
+    line-height: var(--line-height-none);
     white-space: nowrap;
     backdrop-filter: blur(var(--badge-blur, 0));
     -webkit-backdrop-filter: blur(var(--badge-blur, 0));